Updates
Organized asset menu, added mesh for enemy model, added enemy health UI to new model, created a medkit in blender and imported it into unity to act as medkit, all scripts updated to use medkit properly for only the player. Next will be animations for enemyAI and I will attempt to create a new model in blender for melee enemies.